+++ R/format.R
@@ -134,13 +134,16 @@ format.marginaleffects_means <- function(x, model, ci = 0.95, ...) {
   }
   non_focal <- setdiff(colnames(model_data), attr(x, "focal_terms"))
   is_contrast_analysis <- !is.null(list(...)$hypothesis)
+  is_inequality_analysis <- is_contrast_analysis && identical(list(...)$hypothesis, "inequality")
   predict_type <- attributes(x)$predict
 
   # define all columns that should be removed
   remove_columns <- c("s.value", "S", "CI", "rowid_dedup", non_focal)
 
   # do we have contrasts? For contrasts, we want to keep p-values
-  if (is_contrast_analysis) {
+  if (is_inequality_analysis) {
+    estimate_name <- "Mean_Difference"
+  } else if (is_contrast_analysis) {
     estimate_name <- "Difference"
   } else {
     # for simple means, we don't want p-values
@@ -248,9 +251,10 @@ format.marginaleffects_contrasts <- function(x, model = NULL, p_adjust = NULL, c
   }
 
   # for contrasting slopes, we do nothing more here. for other contrasts,
-  # we prettify labels now
+  # we prettify labels now. For special inequality contrasts, we also need no
+  # cleaning, so we skip here, too
 
-  if (!is.null(comparison)) {
+  if (!is.null(comparison) && !identical(comparison, "inequality") && !identical(comparison, "total")) {
     #  the goal here is to create tidy columns with the comparisons.
     # marginaleffects returns a single column that contains all levels that
     # are contrasted. We want to have the contrasted levels per predictor in
@@ -597,6 +601,9 @@ format.marginaleffects_contrasts <- function(x, model = NULL, p_adjust = NULL, c
   # finally, make sure we have original data types
   params <- data.frame(datawizard::data_restoretype(params, model_data))
 
+  # fix for inequality-comparisons
+  colnames(params)[colnames(params) == "Mean_Difference"] <- "Mean Difference"
+
   # add posterior draws?
   if (!is.null(attributes(x)$posterior_draws)) {
     # how many?
@@ -668,7 +675,7 @@ format.marginaleffects_contrasts <- function(x, model = NULL, p_adjust = NULL, c
 #' @keywords internal
 .guess_estimate_name <- function(predict_type, info) {
   # estimate name
-  if (is.null(predict_type) && is.null(info)) {
+  if (is.null(predict_type)) {
     estimate_name <- "Mean"
   } else if (!is.null(predict_type) && tolower(predict_type) %in% .brms_aux_elements()) {
     # for Bayesian models with distributional parameter